+++ src/Repull.SDK/V1/Migrations/Item/ChannelMap/ChannelMapResponse.cs
@@ -5,23 +5,23 @@ using Microsoft.Kiota.Abstractions.Serialization;
 using System.Collections.Generic;
 using System.IO;
 using System;
-namespace Repull.SDK.V1.Health.Atlas
+namespace Repull.SDK.V1.Migrations.Item.ChannelMap
 {
-    [Obsolete("This class is obsolete. Use AtlasGetResponse instead.")]
+    [Obsolete("This class is obsolete. Use ChannelMapGetResponse instead.")]
     [global::System.CodeDom.Compiler.GeneratedCode("Kiota", "1.0.0")]
     #pragma warning disable CS1591
-    public partial class AtlasResponse : global::Repull.SDK.V1.Health.Atlas.AtlasGetResponse, IParsable
+    public partial class ChannelMapResponse : global::Repull.SDK.V1.Migrations.Item.ChannelMap.ChannelMapGetResponse, IParsable
     #pragma warning restore CS1591
     {
         /// <summary>
         /// Creates a new instance of the appropriate class based on discriminator value
         /// </summary>
-        /// <returns>A <see cref="global::Repull.SDK.V1.Health.Atlas.AtlasResponse"/></returns>
+        /// <returns>A <see cref="global::Repull.SDK.V1.Migrations.Item.ChannelMap.ChannelMapResponse"/></returns>
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
-        public static new global::Repull.SDK.V1.Health.Atlas.AtlasResponse CreateFromDiscriminatorValue(IParseNode parseNode)
+        public static new global::Repull.SDK.V1.Migrations.Item.ChannelMap.ChannelMapResponse CreateFromDiscriminatorValue(IParseNode parseNode)
         {
             if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
-            return new global::Repull.SDK.V1.Health.Atlas.AtlasResponse();
+            return new global::Repull.SDK.V1.Migrations.Item.ChannelMap.ChannelMapResponse();
         }
     }
 }