+++ src/Repull.SDK/V1/Migrations/Item/Import/ImportPostResponse.cs
@@ -5,31 +5,39 @@ using Microsoft.Kiota.Abstractions.Serialization;
 using System.Collections.Generic;
 using System.IO;
 using System;
-namespace Repull.SDK.V1.Health.Atlas
+namespace Repull.SDK.V1.Migrations.Item.Import
 {
     [global::System.CodeDom.Compiler.GeneratedCode("Kiota", "1.0.0")]
     #pragma warning disable CS1591
-    public partial class AtlasGetResponse : IAdditionalDataHolder, IParsable
+    public partial class ImportPostResponse : IAdditionalDataHolder, IParsable
     #pragma warning restore CS1591
     {
         /// <summary>Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.</summary>
         public IDictionary<string, object> AdditionalData { get; set; }
+        /// <summary>The data property</summary>
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_1_OR_GREATER
+#nullable enable
+        public global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse_data? Data { get; set; }
+#nullable restore
+#else
+        public global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse_data Data { get; set; }
+#endif
         /// <summary>
-        /// Instantiates a new <see cref="global::Repull.SDK.V1.Health.Atlas.AtlasGetResponse"/> and sets the default values.
+        /// Instantiates a new <see cref="global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse"/> and sets the default values.
         /// </summary>
-        public AtlasGetResponse()
+        public ImportPostResponse()
         {
             AdditionalData = new Dictionary<string, object>();
         }
         /// <summary>
         /// Creates a new instance of the appropriate class based on discriminator value
         /// </summary>
-        /// <returns>A <see cref="global::Repull.SDK.V1.Health.Atlas.AtlasGetResponse"/></returns>
+        /// <returns>A <see cref="global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse"/></returns>
         /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
-        public static global::Repull.SDK.V1.Health.Atlas.AtlasGetResponse CreateFromDiscriminatorValue(IParseNode parseNode)
+        public static global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse CreateFromDiscriminatorValue(IParseNode parseNode)
         {
             if(ReferenceEquals(parseNode, null)) throw new ArgumentNullException(nameof(parseNode));
-            return new global::Repull.SDK.V1.Health.Atlas.AtlasGetResponse();
+            return new global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse();
         }
         /// <summary>
         /// The deserialization information for the current model
@@ -39,6 +47,7 @@ namespace Repull.SDK.V1.Health.Atlas
         {
             return new Dictionary<string, Action<IParseNode>>
             {
+                { "data", n => { Data = n.GetObjectValue<global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse_data>(global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse_data.CreateFromDiscriminatorValue); } },
             };
         }
         /// <summary>
@@ -48,6 +57,7 @@ namespace Repull.SDK.V1.Health.Atlas
         public virtual void Serialize(ISerializationWriter writer)
         {
             if(ReferenceEquals(writer, null)) throw new ArgumentNullException(nameof(writer));
+            writer.WriteObjectValue<global::Repull.SDK.V1.Migrations.Item.Import.ImportPostResponse_data>("data", Data);
             writer.WriteAdditionalData(AdditionalData);
         }
     }